+++ lib/function/index.ts
@@ -1,0 +1,21 @@
+'use strict';
+
+import {isFunction}         from './is';
+import {isAsyncFunction}    from './isAsync';
+import {noop}               from './noop';
+import {noopresolve}        from './noopresolve';
+import {noopreturn}         from './noopreturn';
+import {sleep}              from './sleep';
+
+export {
+    isFunction,
+    isFunction as is,
+    isFunction as isFn,
+    isAsyncFunction,
+    isAsyncFunction as isAsync,
+    isAsyncFunction as isAsyncFn,
+    noop,
+    noopresolve,
+    noopreturn,
+    sleep,
+};